requirements.txt
asgiref==3.3.4
Django==3.2.4
djangorestframework==3.12.4
pkg-resources==0.0.0
psycopg2-binary==2.8.6
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.10.0.0
